1014 17：21整合前版本

+++ Project04/src/main/webapp/js/chat.js
@@ -1,6 +1,55 @@
 ﻿
     var fromid = '';
     var toid = '';
+    var chatState = 'close';
+
+    !function(e, t, a) {
+        function r() {
+            for (var e = 0; e < s.length; e++) s[e].alpha <= 0 ? (t.body.removeChild(s[e].el), s.splice(e, 1)) : (s[e].y--, s[e].scale += .004, s[e].alpha -= .013, s[e].el.style.cssText = "left:" + s[e].x + "px;top:" + s[e].y + "px;opacity:" + s[e].alpha + ";transform:scale(" + s[e].scale + "," + s[e].scale + ") rotate(45deg);background:" + s[e].color + ";z-index:99999");
+            requestAnimationFrame(r)
+        }
+        function n() {
+            var t = "function" == typeof e.onclick && e.onclick;
+            e.onclick = function(e) {
+                t && t(),
+                    o(e)
+            }
+        }
+        function o(e) {
+            var a = t.createElement("div");
+            a.className = "heart",
+                s.push({
+                    el: a,
+                    x: e.clientX - 5,
+                    y: e.clientY - 5,
+                    scale: 1,
+                    alpha: 1,
+                    color: c()
+                }),
+                t.body.appendChild(a)
+        }
+        function i(e) {
+            var a = t.createElement("style");
+            a.type = "text/css";
+            try {
+                a.appendChild(t.createTextNode(e))
+            } catch(t) {
+                a.styleSheet.cssText = e
+            }
+            t.getElementsByTagName("head")[0].appendChild(a)
+        }
+        function c() {
+            return "rgb(" + "240" + "," + "80" + "," + ~~ (255 * Math.random()) + ")"
+        }
+        var s = [];
+        e.requestAnimationFrame = e.requestAnimationFrame || e.webkitRequestAnimationFrame || e.mozRequestAnimationFrame || e.oRequestAnimationFrame || e.msRequestAnimationFrame ||
+            function(e) {
+                setTimeout(e, 1e3 / 60)
+            },
+            i(".heart{width: 10px;height: 10px;position: fixed;background: #f00;transform: rotate(45deg);-webkit-transform: rotate(45deg);-moz-transform: rotate(45deg);}.heart:after,.heart:before{content: '';width: inherit;height: inherit;background: inherit;border-radius: 50%;-webkit-border-radius: 50%;-moz-border-radius: 50%;position: fixed;}.heart:after{top: -5px;}.heart:before{left: -5px;}"),
+            n(),
+            r()
+    } (window, document);
 
     function getParam(userid,url) {
         fromid = userid;
@@ -39,12 +88,26 @@
     $(".chatBtn").click(function () {
         $(".chatBox").toggle(0);
         //获得聊天列表
-        showChatList(fromid);
-        actionControl();
+        if (chatState=='close'){
+            showChatList(fromid);
+            actionControl();
+            chatState = 'list';
+        }
+        else if (chatState=='show') {
+            $(".chat-return").trigger("click");
+            chatState = 'close';
+        }
+        else if (chatState=='list'){
+            chatState = 'close';
+        }
     })
 
     $(".chat-close").click(function () {
         $(".chatBox").toggle(10);
+        if (chatState=='show') {
+            $(".chat-return").trigger("click");
+        }
+        chatState = 'close';
     })
 
     var totalNum = $(".chat-message-num").html();
@@ -69,7 +132,7 @@
         // 进聊天页面
         $(".chat-list-people").each(function () {
             $(this).click(function () {
-                alert('进入聊天页面')
+
                 var offline = $(this).next().val();
                 if (offline=="12"){
                     alert('对方不在线！')
@@ -77,6 +140,8 @@
                 }
 
                 toid = $(this).prev().val();
+
+                $(".chatBox-content-demo").html('');
 
                 var n = $(this).index();
                 $(".chatBox-head-one").toggle();
@@ -87,15 +152,17 @@
                 //传名字
                 $(".ChatInfoName").text($(this).children(".chat-name").children("p").eq(0).html());
 
+                chatState = 'show';
+
                 //加载送礼图标
                 giftText="<div class='gift-send'><a href='javascript:void(0);' title='赠送礼物'><img  src='images/gift.png' alt='赠送礼物'></a></div>";
                 $(".ChatInfoName").append(giftText);
 
                 //赠送礼物
                 $(".gift-send").click(function(){
-                    gifturl = "MyFollowManager/gift.action?toid="+toid;
+
                     if (confirm('将扣除100金币，确认赠送礼物？')){
-                        location.href=gifturl;
+                        giftSend(toid);
                     }
                 });
 
@@ -121,14 +188,16 @@
             $(".chatBox-list").fadeToggle(1);
             $(".chatBox-kuang").fadeToggle(1);
             $(".chatBox-content-demo").html('');
+            chatState = 'list';
         });
 
         //删除对象
         $(".chat-delete").off("click");
         $(".chat-delete").click(function (event) {
-            alert('进入删除逻辑');
-            toid = $(this).parent().prev().val();
-            deleteChatUser(fromid,toid);
+            if (confirm('是否删除？（将清空聊天记录）')){
+                toid = $(this).parent().prev().val();
+                deleteChatUser(fromid,toid);
+            }
             event.stopPropagation();
         });
 
@@ -204,11 +273,6 @@
             return false;
         }
         var uploadResult = uploadFile(pic);
-
-        if (uploadResult=="yes") {
-            //聊天框默认最底部
-
-        }
 
     }
 
@@ -326,6 +390,110 @@
                 alert('删除失败，连接异常');
             }
         });
+    }
+
+    //赠送送礼
+    function giftSend(gtoid) {
+        $.ajax({
+            async : false, //设置同步
+            type : 'POST',
+            url : 'MyFollowManager/gift.action',
+            data : {toid:gtoid},
+            success : function(result){
+                if (result=='yes'){
+                    alert('赠送成功');
+                    yanhua();
+                }
+                else if (result=='noenough') {
+                    alert('您的金币不足');
+                }
+                else if (result=='no'){
+                    alert('赠送礼物失败，未知异常');
+                }
+            },
+            error : function(result){
+                alert('赠送礼物失败，连接异常');
+            }
+        });
+    }
+
+    function yanhua() {
+
+            var aDiv=[];
+            var oDiv=null;
+            var _oDiv=document.createElement('div');
+            var i=0;
+
+            var chatBox = $(".chatBox-content-demo");
+            var x  = (document.body.clientWidth-chatBox.offset().left)/2+chatBox.offset().left;
+            var y = (document.body.clientHeight-chatBox.offset().top)/2+chatBox.offset().top;
+
+            _oDiv.style.position='absolute';
+            _oDiv.style.background='red';
+            _oDiv.style.width='3px';
+            _oDiv.style.height='30px';
+            _oDiv.style.left=x+'px';
+            _oDiv.style.top=document.documentElement.clientHeight-30+'px';
+
+            document.body.appendChild(_oDiv);
+
+            var t=setInterval(function (){
+                if(_oDiv.offsetTop<=y)
+                {
+                    clearInterval(t);
+                    show();
+                    document.body.removeChild(_oDiv);
+                }
+                _oDiv.style.top=_oDiv.offsetTop-30+'px';
+            }, 30);
+
+            function show()
+            {
+                var oDiv=null;
+
+                for(i=0;i<100;i++)
+                {
+                    oDiv=document.createElement('div');
+
+                    oDiv.style.width='3px';
+                    oDiv.style.height='3px';
+                    oDiv.style.background='#'+Math.ceil(Math.random()*0xEFFFFF+0x0FFFFF).toString(16);
+                    oDiv.style.position='absolute';
+                    oDiv.style.left=x+'px';
+                    oDiv.style.top=y-30+'px';
+
+                    var a=Math.random()*360;
+
+                    //oDiv.speedX=Math.random()*40-20;
+                    //oDiv.speedY=Math.random()*40-20;
+
+                    oDiv.speedX=Math.sin(a*180/Math.PI)*20*Math.random();
+                    oDiv.speedY=Math.cos(a*180/Math.PI)*20*Math.random();
+
+                    document.body.appendChild(oDiv);
+
+                    aDiv.push(oDiv);
+                }
+            }
+
+            setInterval(doMove, 30);
+
+            function doMove()
+            {
+                for(i=0;i<aDiv.length;i++)
+                {
+                    aDiv[i].style.left=aDiv[i].offsetLeft+aDiv[i].speedX+'px';
+                    aDiv[i].style.top=aDiv[i].offsetTop+aDiv[i].speedY+'px';
+                    aDiv[i].speedY+=1;
+
+                    if(aDiv[i].offsetLeft<0 || aDiv[i].offsetLeft>document.documentElement.clientWidth || aDiv[i].offsetTop<0 || aDiv[i].offsetTop>document.documentElement.clientHeight)
+                    {
+                        document.body.removeChild(aDiv[i]);
+                        aDiv.splice(i, 1);
+                    }
+                }
+            }
+
     }
